+++ tests/fixtures/runtime_errors/interface_exists_invalid_autoload.php
@@ -1,0 +1,2 @@
+<?php
+var_dump(interface_exists("Box", 1));